review fixes

+++ contracts/priceOracle/IChronicle.sol
@@ -10,7 +10,7 @@ interface IChronicle {
     /// @notice Returns the oracle's current value.
     /// @dev Reverts if no value set.
     /// @return value The oracle's current value.
-    function read() external view returns (uint value);
+    function read() external view returns (uint256 value);
 }
 
 interface IToll {
